Source/SDK/PayPal/Api/Payments/Phone.cs
using Newtonsoft.Json;

namespace PayPal.Api.Payments
{
    public class Phone
    {
        /// <summary>
        /// Country code (from in E.164 format)
        /// </summary>
        [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "country_code")]
        public string country_code { get; set; }

        /// <summary>
        /// In-country phone number (from in E.164 format)
        /// </summary>
        [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "number")]
        public string national_number { get; set; }

        /// <summary>
        /// Phone extension
        /// </summary>
        [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "extension")]
        public string extension { get; set; }

        /// <summary>
        /// Converts the object to JSON string
        /// </summary>
        public virtual string ConvertToJson()
        {
            return JsonFormatter.ConvertToJson(this);
        }
    }
}
